ultracache/tests/requirements/16.txt
django==1.6.11
requests==2.10.0
djangorestframework==3.3.0
